+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v0.2.0
+
+ * Added stats.js performance monitoring.
+ * Added support for passes other than shaders.
+ * Improved props mapping from constructor properties.
+ * Added WebGLDeferredRenderer as a renderer option.
+ * Fixed incorrect dependency usage in Webpack.
+
 ### v0.1.0
 
  * Initial commit.